+++ conn_test.go
@@ -862,19 +862,17 @@ func TestClientNewSessionInvalidSecondResponseDifferentChannel(t *testing.T) {
 }
 
 func TestNewSessionTimedOut(t *testing.T) {
-	endAck := make(chan struct{})
 	responder := func(req frames.FrameBody) ([]byte, error) {
 		switch req.(type) {
 		case *fake.AMQPProto:
 			return []byte{'A', 'M', 'Q', 'P', 0, 1, 0, 0}, nil
 		case *frames.PerformOpen:
 			return fake.PerformOpen("container")
+		case *frames.PerformClose:
+			return fake.PerformClose(nil)
 		case *frames.PerformBegin:
 			// swallow the frame so NewSession never gets an ack
 			return nil, nil
-		case *frames.PerformEnd:
-			close(endAck)
-			return fake.PerformEnd(0, nil)
 		default:
 			return nil, fmt.Errorf("unhandled frame %T", req)
 		}
@@ -891,13 +889,6 @@ func TestNewSessionTimedOut(t *testing.T) {
 	cancel()
 	require.ErrorIs(t, err, context.DeadlineExceeded)
 	require.Nil(t, session)
-
-	select {
-	case <-time.After(time.Second):
-		t.Fatal("didn't receive end ack")
-	case <-endAck:
-		// expected
-	}
 }
 
 func TestNewSessionWriteError(t *testing.T) {
@@ -937,45 +928,5 @@ func TestNewSessionWriteError(t *testing.T) {
 		// expected
 	case <-endAck:
 		t.Fatal("unexpected ack")
-	}
-}
-
-func TestNewSessionTimedOutAckTimedOut(t *testing.T) {
-	endAck := make(chan struct{})
-	responder := func(req frames.FrameBody) ([]byte, error) {
-		switch req.(type) {
-		case *fake.AMQPProto:
-			return []byte{'A', 'M', 'Q', 'P', 0, 1, 0, 0}, nil
-		case *frames.PerformOpen:
-			return fake.PerformOpen("container")
-		case *frames.PerformBegin:
-			// swallow the frame so NewSession never gets an ack
-			return nil, nil
-		case *frames.PerformEnd:
-			close(endAck)
-			// swallow the frame so the closing goroutine never gets an ack
-			return nil, nil
-		default:
-			return nil, fmt.Errorf("unhandled frame %T", req)
-		}
-	}
-	netConn := fake.NewNetConn(responder)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-	client, err := NewConn(ctx, netConn, nil)
-	cancel()
-	require.NoError(t, err)
-	// fisrt session succeeds
-	ctx, cancel = context.WithTimeout(context.Background(), 1*time.Second)
-	session, err := client.NewSession(ctx, nil)
-	cancel()
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	require.Nil(t, session)
-
-	select {
-	case <-time.After(time.Second):
-		t.Fatal("didn't receive end ack")
-	case <-endAck:
-		// expected
 	}
 }